Create Get-MountPointData.ps1
This script collects mount point info from remote windows hosts, and also the EMC LUN info. This works correctly only if the remote host has PowerMT installed. If not, then it will give an error message, but the collected info will be OK.
I tested only if there is one volume (partition) created on a physical device. There could be issues if the configuration is not simple like that, but currently I don't have a chance for testing.