For pushing the current scripts
Get-LUNIDfromNAA - It is not my script, downloaded, but I want to have it handy

Get-ServerHBAPorts - Basically a "get-initiatorport", but then the wwn is converted to a format that is useful in switches

isilonReport - pulls and stores report from isilon

switchCheck - simply just checks if a configured wwn is logged in or not

wastedStorageGroupList - In case you use thin provisioning on vmax, you can check the percentage used in the LUNs. the storage array serial and the percentage should be passed to the script.